Add information on libraries used by Anaconda & Canopy
external-so.sh is a little script to find libraries that are required by
files in some directory but are not provided by any files in that
directory.

{anaconda,canopy}-*-libraries are the output of running this script on
the specified distributions.

The canopy results are kinda weird and implausible, so to aid discussion
there's also a file in canopy-further-investigation/ listing all the ELF
files in canopy and what they link to.