Disable Travis "branch builds" for pyup- branches
pyup.io works by creating a branch in the main repo, and then creating
a PR from that branch to master. Travis therefore ends up testing
these commits twice: once because they're a PR, and once because
they're on a branch in the main repo.

This should hopefully disable the branch-based testing, so that we end
up with each PR being tested only once.