Build hardening (#1009)
* Use hardening for building all tools & libraries

This does not affect the wheels that are produced by end users as proposed in https://github.com/pypa/manylinux/issues/59 but mitigates potential security issues in the tools used by manylinux images as mentioned in https://github.com/pypa/manylinux/issues/1005

* Always update system packages in the final step

Since docker cache is used, system packages are not updated when cache is present. Always update them in the final step.