Do not break /usr/local layout
Some distros are defining /usr/local/man to be a symlink to /usr/local/share/man

The clean-up done in update-system-packages.sh was a bit too aggressive and lead to breaking the symlink.
Only delete the content of /usr/local/share/man not to break the layout as defined by the distros.

Fix #1060